move oracle related functionality to oracle_provider.go
