+++ client/api.go
@@ -1,22 +1,16 @@
 package client
 
 import (
-	"crypto/rsa"
 	"fmt"
-	"io/ioutil"
 	"net/url"
 	"os"
-	"path/filepath"
 	"strings"
-	"syscall"
 
 	"github.com/fnproject/cli/config"
 	fnclient "github.com/fnproject/fn_go/client"
 	openapi "github.com/go-openapi/runtime/client"
 	"github.com/go-openapi/strfmt"
-	oci "github.com/oracle/oci-go-sdk/common"
 	"github.com/spf13/viper"
-	"golang.org/x/crypto/ssh/terminal"
 )
 
 func Host() string {
@@ -63,56 +57,15 @@ func defaultProvider(transport *openapi.Runtime) {
 	}
 }
 
-func challengeForPKeyPassword() string {
-	fmt.Print("Private Key Phrase: ")
-	bytePassword, err := terminal.ReadPassword(int(syscall.Stdin))
-	if err != nil {
-		panic(fmt.Sprintf("%s", err))
-	}
-	password := string(bytePassword)
-	fmt.Println()
-
-	return password
-}
-
-func privateKey(pkeyFilePath string) (*rsa.PrivateKey, error) {
-	keyBytes, err := ioutil.ReadFile(pkeyFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to load private key from file: %s. Error: %s \n", pkeyFilePath, err)
-	}
-
-	var key *rsa.PrivateKey
-	pKeyPword := viper.GetString(config.OraclePassPhrase)
-
-	if !viper.IsSet(config.OraclePassPhrase) {
-		if key, err = getPrivateKey(keyBytes, pKeyPword, pkeyFilePath); key == nil {
-			pKeyPword = challengeForPKeyPassword()
-		}
-	}
-	key, err = getPrivateKey(keyBytes, pKeyPword, pkeyFilePath)
-	return key, err
-}
-
-func getPrivateKey(keyBytes []byte, pKeyPword, pkeyFilePath string) (*rsa.PrivateKey, error) {
-	key, err := oci.PrivateKeyFromBytes(keyBytes, oci.String(pKeyPword))
-	if err != nil {
-		if pKeyPword != "" {
-			return nil, fmt.Errorf("Unable to load private key from file bytes: %s. Error: %s \n", pkeyFilePath, err)
-		}
-	}
-
-	return key, nil
-}
-
 func oracleProvider(transport *openapi.Runtime) error {
-	keyID, pKey, err := oracleConfigFile()
+	keyID, pKey, err := OracleConfigFile()
 	if err != nil {
 		return err
 	}
 
-	compartmentID := viper.GetString(config.OracleCompartmentID)
+	compartmentID := viper.GetString(oracleCompartmentID)
 
-	if viper.GetBool(config.OracleDisableCerts) {
+	if viper.GetBool(oracleDisableCerts) {
 		transport.Transport = InsecureRoundTripper(transport.Transport)
 	}
 
@@ -125,60 +78,6 @@ func oracleProvider(transport *openapi.Runtime) error {
 				transport.Transport))
 
 	return err
-}
-
-func oracleConfigFile() (string, *rsa.PrivateKey, error) {
-	var oracleProfile string
-	if oracleProfile = viper.GetString(config.OracleProfile); oracleProfile == "" {
-		oracleProfile = "DEFAULT"
-	}
-
-	cf, err := oci.ConfigurationProviderFromFileWithProfile(filepath.Join(config.GetHomeDir(), ".oci", "config"), oracleProfile, "")
-	if err != nil {
-		return "", nil, err
-	}
-
-	var tenancyID string
-	if tenancyID = viper.GetString(config.OracleTenancyID); tenancyID == "" {
-		tenancyID, err = cf.TenancyOCID()
-		if err != nil {
-			return "", nil, err
-		}
-	}
-
-	var userID string
-	if userID = viper.GetString(config.OracleUserID); userID == "" {
-		userID, err = cf.UserOCID()
-		if err != nil {
-			return "", nil, err
-		}
-	}
-
-	var fingerprint string
-	if fingerprint = viper.GetString(config.OracleFingerprint); fingerprint == "" {
-		fingerprint, err = cf.KeyFingerprint()
-		if err != nil {
-			return "", nil, err
-		}
-	}
-
-	keyID := tenancyID + "/" + userID + "/" + fingerprint
-	var pKey *rsa.PrivateKey
-	if keyFile := viper.GetString(config.OracleKeyFile); keyFile != "" {
-		pKey, err = privateKey(keyFile)
-		if err != nil {
-			return "", nil, err
-		}
-		return keyID, pKey, nil
-	}
-
-	// Read private key for .oci file
-	pKey, err = cf.PrivateRSAKey()
-	if err != nil {
-		return "", nil, err
-	}
-
-	return keyID, pKey, nil
 }
 
 func GetTransportAndRegistry() (*openapi.Runtime, strfmt.Registry, error) {
